refactor: move track matching from hit to cluster level
