+++ reconstruction/alert/src/main/java/org/jlab/rec/atof/Hit/HitFinder.java
@@ -5,7 +5,6 @@ import java.util.Collections;
 import org.jlab.geom.base.Detector;
 import org.jlab.io.base.DataBank;
 import org.jlab.io.base.DataEvent;
-import org.jlab.rec.atof.trackMatch.TrackProjector;
 
 /**
  * The {@code HitFinder} class finds hits in the atof.
@@ -64,10 +63,8 @@ public class HitFinder {
      * @param event the {@link DataEvent} containing hits.
      * @param atof the {@link Detector} representing the atof geometry to match
      * the sector/layer/component to x/y/z.
-     * @param track_projector the {@link TrackProjector} containing the ahdc tracks projected
-     * to the atof for matching.
      */
-    public void findHits(DataEvent event, Detector atof, TrackProjector track_projector) {
+    public void findHits(DataEvent event, Detector atof) {
         //For each event a list of bar hits and a list of wedge hits are filled
         this.barHits.clear();
         this.wedgeHits.clear();
@@ -103,7 +100,6 @@ public class HitFinder {
                     case "bar down" ->
                         hit_down.add(hit);
                     case "wedge" -> {
-                        hit.matchTrack(track_projector);
                         this.wedgeHits.add(hit);
                     }
                     default ->
@@ -122,80 +118,6 @@ public class HitFinder {
                 if (this_hit_up.matchBar(this_hit_down)) {
                     //Bar hits are matched to ahdc tracks and listed
                     BarHit this_bar_hit = new BarHit(this_hit_up, this_hit_down);
-                    this_bar_hit.matchTrack(track_projector);
-                    this.barHits.add(this_bar_hit);
-                }
-            }
-        }
-        //Once all has been listed, hits are sorted by energy
-        Collections.sort(this.barHits, (hit1, hit2) -> Double.compare(hit2.getEnergy(), hit1.getEnergy()));
-        Collections.sort(this.wedgeHits, (hit1, hit2) -> Double.compare(hit2.getEnergy(), hit1.getEnergy()));
-    }
-
-    /**
-     * Find hits in the event, matches them to tracks in the 
-     * projections bank and build their properties.
-     *
-     * @param event the {@link DataEvent} containing hits and the bank with ahdc track projections.
-     * @param atof the {@link Detector} representing the atof geometry to match
-     * the sector/layer/component to x/y/z.
-     */
-    public void findHits(DataEvent event, Detector atof) {
-
-        //For each event a list of bar hits and a list of wedge hits are filled
-        this.barHits.clear();
-        this.wedgeHits.clear();
-        //They are read from the ATOF TDC bank
-        DataBank bank_atof_hits = event.getBank("ATOF::tdc");
-        int nt = bank_atof_hits.rows(); // number of hits
-        //Hits in the bar downstream and upstream will be matched
-        ArrayList<AtofHit> hit_up = new ArrayList<>();
-        ArrayList<AtofHit> hit_down = new ArrayList<>();
-        //Looping through all hits
-        for (int i = 0; i < nt; i++) {
-            //Getting their properties
-            int sector = bank_atof_hits.getInt("sector", i);
-            int layer = bank_atof_hits.getInt("layer", i);
-            int component = bank_atof_hits.getInt("component", i);
-            int order = bank_atof_hits.getInt("order", i);
-            int tdc = bank_atof_hits.getInt("TDC", i);
-            int tot = bank_atof_hits.getInt("ToT", i);
-            //Building a Hit
-            AtofHit hit = new AtofHit(sector, layer, component, order, tdc, tot, atof);
-            if (hit.getEnergy() < 0.01) {
-                continue; //energy threshold
-            }                //Sorting the hits into wedge, upstream and downstream bar hits
-            //Lists are built for up/down bar to match them after
-            //Wedge hits are mayched to ahdc tracks and listed 
-            if (null == hit.getType()) {
-                System.out.print("Undefined hit type \n");
-            } else {
-                switch (hit.getType()) {
-                    case "bar up" ->
-                        hit_up.add(hit);
-                    case "bar down" ->
-                        hit_down.add(hit);
-                    case "wedge" -> {
-                        hit.matchTrack(event);
-                        this.wedgeHits.add(hit);
-                    }
-                    default ->
-                        System.out.print("Undefined hit type \n");
-                }
-            }
-        }//End loop through all hits
-
-        //Starting loop through up hits in the bar
-        for (int i_up = 0; i_up < hit_up.size(); i_up++) {
-            AtofHit this_hit_up = hit_up.get(i_up);
-            //Starting loop through down hits in the bar
-            for (int i_down = 0; i_down < hit_down.size(); i_down++) {
-                AtofHit this_hit_down = hit_down.get(i_down);
-                //Matching the hits: if same module and different order, they make up a bar hit
-                if (this_hit_up.matchBar(this_hit_down)) {
-                    //Bar hits are matched to ahdc tracks and listed
-                    BarHit this_bar_hit = new BarHit(this_hit_up, this_hit_down);
-                    //this_bar_hit.matchTrack(event);
                     this.barHits.add(this_bar_hit);
                 }
             }
